requirements.txt
bcrypt==4.0.1
blinker==1.7.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.3
pycparser==2.21
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
typing_extensions==4.8.0
Werkzeug==3.0.1
bcrypt==4.0.1
blinker==1.7.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.3
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
typing_extensions==4.8.0
Werkzeug==3.0.1
